+++ src/runtime/components/Map.vue
@@ -1,17 +1,128 @@
 <!-- eslint-disable vue/multi-word-component-names -->
 <script setup lang="ts">
-    import { MapboxOptions } from 'mapbox-gl';
+    import { EventData, MapboxEvent, MapboxOptions, MapBoxZoomEvent, MapDataEvent, MapMouseEvent, MapTouchEvent, MapWheelEvent } from 'mapbox-gl';
     import { provide, onMounted, StyleValue } from 'vue';
-    import { _mapboxInit } from '../composables/mapboxInit';
+    import { defineMapboxInstance } from '../composables/defineMapboxInstance';
     
     type MapboxComponentOptions = Omit<MapboxOptions, "container">
 
     const props = defineProps<{ mapId: string, options?: MapboxComponentOptions }>()
 
+    const emit = defineEmits<{
+      (e: 'resize'): void
+      (e: 'remove'): void
+
+      (e: 'load'): void
+      (e: 'render'): void
+      (e: 'idle'): void
+      (e: 'error'): void
+      (e: 'webglcontextlost'): void
+      (e: 'webglcontextrestored'): void
+
+      (e: 'data', event: MapDataEvent): void
+      (e: 'styledata', event: MapDataEvent): void
+      (e: 'sourcedata', event: MapDataEvent): void
+      (e: 'dataloading', event: MapDataEvent): void
+      (e: 'styledataloading', event: MapDataEvent): void
+      (e: 'sourcedataloading', event: MapDataEvent): void
+      (e: 'styleimagemissing', event: MapDataEvent): void
+
+      (e: 'movestart', event: MapboxEvent<MouseEvent | TouchEvent | WheelEvent | undefined> & EventData): void
+      (e: 'move', event: MapboxEvent<MouseEvent | TouchEvent | WheelEvent | undefined> & EventData): void
+      (e: 'moveend', event:  MapboxEvent<MouseEvent | TouchEvent | WheelEvent | undefined> & EventData): void
+      (e: 'drag', event:  MapboxEvent<MouseEvent | TouchEvent | undefined> & EventData | MapTouchEvent): void
+      (e: 'dragend', event: MapboxEvent<MouseEvent | TouchEvent | undefined> & EventData | MapTouchEvent): void
+      (e: 'zoomstart', event: MapboxEvent<MouseEvent | TouchEvent | undefined> & EventData | MapTouchEvent): void
+      (e: 'zoom', event: MapboxEvent<MouseEvent | TouchEvent | undefined> & EventData | MapTouchEvent): void
+      (e: 'zoomend', event: MapboxEvent<MouseEvent | TouchEvent | undefined> & EventData | MapTouchEvent): void
+      (e: 'rotatestart', event: MapboxEvent<MouseEvent | TouchEvent | undefined> & EventData | MapTouchEvent): void
+      (e: 'rotate', event: MapboxEvent<MouseEvent | TouchEvent | undefined> & EventData | MapTouchEvent): void
+      (e: 'rotateend', event: MapboxEvent<MouseEvent | TouchEvent | undefined> & EventData | MapTouchEvent): void
+      (e: 'pitchstart', event: MapboxEvent<MouseEvent | TouchEvent | undefined> & EventData | MapTouchEvent): void
+      (e: 'pitch', event: MapboxEvent<MouseEvent | TouchEvent | undefined> & EventData | MapTouchEvent): void
+      (e: 'pitchend', event: MapboxEvent<MouseEvent | TouchEvent | undefined> & EventData | MapTouchEvent): void
+      (e: 'boxzoomstart', event: MapBoxZoomEvent): void
+      (e: 'boxzoomend', event: MapBoxZoomEvent): void
+      (e: 'boxzoomcancel', event: MapBoxZoomEvent): void
+
+      (e: 'mousedown', event: MapMouseEvent): void
+      (e: 'mouseup', event: MapMouseEvent): void
+      (e: 'mouseover', event: MapMouseEvent): void
+      (e: 'mousemove', event: MapMouseEvent): void
+      (e: 'preclick', event: MapMouseEvent): void
+      (e: 'click', event: MapMouseEvent): void
+      (e: 'dblclick', event: MapMouseEvent): void
+      (e: 'mouseenter', event: MapMouseEvent): void
+      (e: 'mouseleave', event: MapMouseEvent): void
+      (e: 'mouseout', event: MapMouseEvent): void
+      (e: 'contextmenu', event: MapMouseEvent): void
+      (e: 'wheel', event: MapWheelEvent): void
+      (e: 'touchstart', event: MapTouchEvent): void
+      (e: 'touchend', event: MapTouchEvent): void
+      (e: 'touchmove', event: MapTouchEvent): void
+      (e: 'touchcancel', event: MapTouchEvent): void
+
+    }>()
+
     provide('MapID', props.mapId)
     onMounted(() => {
-        _mapboxInit(props.mapId, {...props.options, container: props.mapId})
+      const map = defineMapboxInstance(props.mapId, {...props.options, container: props.mapId});
+
+      map.value?.on('resize', () => { emit("resize") })
+      map.value?.on('remove', () => { emit("resize") })
+
+      map.value?.on('load', () => { emit("load") })
+      map.value?.on('render', () => { emit("render") })
+      map.value?.on('idle', () => { emit("idle") })
+      map.value?.on('error', () => { emit("error") })
+      map.value?.on('webglcontextlost', () => { emit("webglcontextlost") })
+      map.value?.on('webglcontextrestored', () => { emit("webglcontextrestored") })
+
+      map.value?.on('data', (e) => { emit("data", e) })
+      map.value?.on('styledata', (e) => { emit("styledata", e) })
+      map.value?.on('sourcedata', (e) => { emit("sourcedata", e) })
+      map.value?.on('dataloading', (e) => { emit("dataloading", e) })
+      map.value?.on('styledataloading', (e) => { emit("styledataloading", e) })
+      map.value?.on('sourcedataloading', (e) => { emit("sourcedataloading", e) })
+      map.value?.on('styleimagemissing', (e) => { emit("styleimagemissing", e) })
+      
+      map.value?.on('movestart', (e) => { emit("movestart", e) })
+      map.value?.on('move', (e) => { emit("move", e) })
+      map.value?.on('moveend', (e) => { emit("moveend", e) })
+      map.value?.on('drag', (e) => { emit("drag", e) })
+      map.value?.on('dragend', (e) => { emit("dragend", e) })
+      map.value?.on('zoomstart', (e) => { emit("zoomstart", e) })
+      map.value?.on('zoom', (e) => { emit("zoom", e) })
+      map.value?.on('zoomend', (e) => { emit("zoomend", e) })
+      map.value?.on('rotatestart', (e) => { emit("rotatestart", e) })
+      map.value?.on('rotate', (e) => { emit("rotate", e) })
+      map.value?.on('rotateend', (e) => { emit("rotateend", e) })
+      map.value?.on('pitchstart', (e) => { emit("pitchstart", e) })
+      map.value?.on('pitch', (e) => { emit("pitch", e) })
+      map.value?.on('pitchend', (e) => { emit("pitchend", e) })
+      map.value?.on('boxzoomstart', (e) => { emit("boxzoomstart", e) })
+      map.value?.on('boxzoomend', (e) => { emit("boxzoomend", e) })
+      map.value?.on('boxzoomcancel', (e) => { emit("boxzoomcancel", e) })
+
+      map.value?.on('mousedown', (e) => { emit("mousedown", e) })
+      map.value?.on('mouseup', (e) => { emit("mouseup", e) })
+      map.value?.on('mouseover', (e) => { emit("mouseover", e) })
+      map.value?.on('mousemove', (e) => { emit("mousemove", e) })
+      map.value?.on('preclick', (e) => { emit("preclick", e) })
+      map.value?.on('click', (e) => { emit("click", e) })
+      map.value?.on('dblclick', (e) => { emit("dblclick", e) })
+      map.value?.on('mouseenter', (e) => { emit("mouseenter", e) })
+      map.value?.on('mouseleave', (e) => { emit("mouseleave", e) })
+      map.value?.on('mouseout', (e) => { emit("mouseout", e) })
+      map.value?.on('contextmenu', (e) => { emit("contextmenu", e) })
+      map.value?.on('wheel', (e) => { emit("wheel", e) })
+      map.value?.on('touchstart', (e) => { emit("touchstart", e) })
+      map.value?.on('touchend', (e) => { emit("touchend", e) })
+      map.value?.on('touchmove', (e) => { emit("touchmove", e) })
+      map.value?.on('touchcancel', (e) => { emit("touchcancel", e) })
     })
+
+    
 </script>