+++ src/runtime/plugin.client.ts
@@ -21,8 +21,15 @@ export default defineNuxtPlugin((nuxtApp) => {
 
     const appConfig = useAppConfig() as ExtendedAppConfig
     //@ts-ignore
-    mapboxgl.accessToken = appConfig._MAPBOX_API_KEY;
-
+    mapboxgl.accessToken = appConfig._MAPBOX_CONFIG.accessToken;
+    //@ts-ignore
+    if (appConfig._MAPBOX_CONFIG.baseApiUrl) mapboxgl.baseApiUrl = appConfig._MAPBOX_CONFIG.baseApiUrl;
+    //@ts-ignore
+    if (appConfig._MAPBOX_CONFIG.workerUrl) mapboxgl.workerUrl = appConfig._MAPBOX_CONFIG.workerUrl;
+    //@ts-ignore
+    if (appConfig._MAPBOX_CONFIG.workerCount) mapboxgl.workerCount = appConfig._MAPBOX_CONFIG.workerCount;
+    //@ts-ignore
+    if (appConfig._MAPBOX_CONFIG.prewarm) mapboxgl.prewarm();
 
     return {
       provide: {