+++ CHANGELOG.md
@@ -5,9 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## v1.2.0
+## Changed
+- `apiKey` renamed to `accessToken` to match mapbox
+- Renamed _useMapboxInstance to useMapboxInstance
 
-- More config options
+## Added
+- Extended config options
+- Component events from mapbox
+- Exposed defineMapboxInstance for custom map components
+
 ## v1.1.7
 ## Fixed
 - Fixed `Map markers are disabled.` with geocoder